- moved RelationHelpers from smg.Helpers namespace to smg namespace. - Re-added "using System.Linq" statements. - Removed redundant temporary generated .cs files.
