+++ smg/smg.common/StateDescription/Attributes/StateGroupAttribute.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Linq;
 
 namespace smg.Common.StateDescription.Attributes
 {
