+++ smg/smg/StateMachineGenerator.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Linq;
 using System.CodeDom;
 using System.CodeDom.Compiler;
 using System.IO;
