+++ smg/smg/smg.csproj
@@ -50,7 +50,7 @@
     <Compile Include="ArgumentParsing\Exceptions\ArgumentParsingException.cs" />
     <Compile Include="ArgumentParsing\Exceptions\BooleanArgumentPropertyIsNotBoolException.cs" />
     <Compile Include="ExtensionMethods\EnumerableExtensions.cs" />
-    <Compile Include="Helpers\RelationHelpers.cs" />
+    <Compile Include="RelationHelpers.cs" />
     <Compile Include="StateGeneration\DecoratorMethod.cs" />
     <Compile Include="StateGeneration\ExtensionMethods\TypeExtensions.cs" />
     <Compile Include="StateGeneration\StateGenerator.cs" />
